Refactoring: GeoDataRiverSurvey::ChangeSelectionCommand is moved to separated files.
