+++ libs/geodata/riversurvey/geodatariversurvey.cpp
@@ -12,6 +12,7 @@
 #include "geodatariversurveycrosssectionwindowprojectdataitem.h"
 #include "geodatariversurveydisplaysettingdialog.h"
 #include "geodatariversurveyproxy.h"
+#include "private/geodatariversurvey_changeselectioncommand.h"
 #include "private/geodatariversurvey_setdisplaysettingcommand.h"
 
 #include <guicore/pre/base/preprocessorwindowinterface.h>
@@ -1048,75 +1049,6 @@ void GeoDataRiverSurvey::mousePressEvent(QMouseEvent* event, PreProcessorGraphic
 		m_dragStartPoint = QPoint(event->x(), event->y());
 	}
 }
-
-class GeoDataRiverSurvey::ChangeSelectionCommand : public QUndoCommand
-{
-public:
-	ChangeSelectionCommand(GeoDataRiverSurvey* rs, MouseBoundingBox* box) :
-		QUndoCommand {GeoDataRiverSurvey::tr("Selection Change")}
-	{
-		m_rs = rs;
-		// store old selection info.
-		buildSelectedPointsSet(m_oldSelectedPoints);
-
-		// now, update the selection statuses of river path points.
-		double point[3];
-		vtkPoints* points = box->vtkGrid()->GetPoints();
-		QVector2D leftTop, rightTop, leftBottom;
-		// left top
-		points->GetPoint(0, point);
-		leftTop = QVector2D(point[0], point[1]);
-		// right top
-		points->GetPoint(1, point);
-		rightTop = QVector2D(point[0], point[1]);
-		// left bottom
-		points->GetPoint(3, point);
-		leftBottom = QVector2D(point[0], point[1]);
-
-		// do the selection!
-		m_rs->headPoint()->selectRegion(leftTop, rightTop - leftTop, leftBottom - leftTop);
-
-		// store new selection info.
-		buildSelectedPointsSet(m_newSelectedPoints);
-	}
-	void buildSelectedPointsSet(QSet<GeoDataRiverPathPoint*>& set) {
-		set.clear();
-		GeoDataRiverPathPoint* p = m_rs->headPoint();
-		if (p != nullptr) {p = p->nextPoint();}
-		while (p != nullptr) {
-			if (p->IsSelected) {
-				set.insert(p);
-			}
-			p = p->nextPoint();
-		}
-	}
-	void redo() {
-		GeoDataRiverPathPoint* p = m_rs->headPoint();
-		if (p != nullptr) {p = p->nextPoint();}
-		while (p != nullptr) {
-			p->IsSelected = m_newSelectedPoints.contains(p);
-			p = p->nextPoint();
-		}
-		m_rs->updateActionStatus();
-		m_rs->updateSelectionShapeData();
-		m_rs->renderGraphicsView();
-	}
-	void undo() {
-		GeoDataRiverPathPoint* p = m_rs->headPoint();
-		if (p != nullptr) {p = p->nextPoint();}
-		while (p != nullptr) {
-			p->IsSelected = m_oldSelectedPoints.contains(p);
-			p = p->nextPoint();
-		}
-		m_rs->updateActionStatus();
-		m_rs->updateSelectionShapeData();
-		m_rs->renderGraphicsView();
-	}
-private:
-	QSet<GeoDataRiverPathPoint*> m_oldSelectedPoints;
-	QSet<GeoDataRiverPathPoint*> m_newSelectedPoints;
-	GeoDataRiverSurvey* m_rs;
-};
 
 void GeoDataRiverSurvey::mouseReleaseEvent(QMouseEvent* event, PreProcessorGraphicsViewInterface* v)
 {
